requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
google-cloud-storage==2.7.0
SQLAlchemy==1.4.23
cloud-sql-python-connector==1.2.0
mysql-connector-python==8.0.26
pymysql==1.0.2
python-dotenv==0.19.0
gunicorn==20.1.0
requests==2.28.1
uuid==1.30